Update global.css
I noticed that the text wasn't sizing properly when switching from desktop to mobile screen width, so I changed the "px" size to "em" size and it fixed it. Honestly, at this point I'm just Googling, pasting, trying, and failing. BUT, it's coming along! This is why I use WordPress with visual editors, LOL.