Update index.html
Made some big edits. Decided that I didn't want it to be a one-page website (because it looked really ugly). Also decided that the stupid circular corner menu I had was terrible and got rid of it. It has been replaced with a munch nicer menu that fits the theme better. I'm now trying to add some SVG Social Icons into the footer and have them link to my accounts. It's not going well, haha. I've given up for tonight.